Refactor and Move methods into refactor rb file
